+++ lib/views/view_accounts/view_accounts.dart
@@ -158,8 +158,11 @@ class ViewAccountsState extends ViewWidgetState {
   }
 
   @override
-  Widget getPanelForChart(final List<int> indices) {
-    return _getSubViewContentForChart(indices);
+  Widget getPanelForChart({
+    required final List<int> selectedIds,
+    required final bool showAsNativeCurrency,
+  }) {
+    return _getSubViewContentForChart(selectedIds: selectedIds, showAsNativeCurrency: showAsNativeCurrency);
   }
 
   @override