+++ lib/views/view_rentals/view_rentals.dart
@@ -54,8 +54,11 @@ class ViewRentalsState extends ViewWidgetState {
   }
 
   @override
-  Widget getPanelForChart(final List<int> indices) {
-    return _getSubViewContentForChart(indices);
+  Widget getPanelForChart({
+    required final List<int> selectedIds,
+    required final bool showAsNativeCurrency,
+  }) {
+    return _getSubViewContentForChart(selectedIds: selectedIds, showAsNativeCurrency: showAsNativeCurrency);
   }
 
   @override