+++ lib/main.dart
@@ -21,7 +21,7 @@ import 'package:money/views/view_accounts/view_accounts.dart';
 import 'package:money/views/view_aliases/view_aliases.dart';
 import 'package:money/views/view_cashflow/view_cashflow.dart';
 import 'package:money/views/view_categories/view_categories.dart';
-import 'package:money/views/view_investments/view_rentals.dart';
+import 'package:money/views/view_investments/view_investments.dart';
 import 'package:money/views/view_loans/view_loans.dart';
 import 'package:money/views/view_payees/view_payees.dart';
 import 'package:money/views/view_rentals/view_rentals.dart';