+++ lib/models/money_objects/stock_splits/stock_splits.dart
@@ -1,4 +1,5 @@
 import 'package:money/helpers/json_helper.dart';
+import 'package:money/models/money_objects/securities/security.dart';
 import 'package:money/models/money_objects/stock_splits/stock_split.dart';
 import 'package:money/models/money_objects/money_objects.dart';
 
@@ -23,5 +24,19 @@ class StockSplits extends MoneyObjects<StockSplit> {
     return super.getCsvFromList(
       getListSortedById(),
     );
+  }
+
+  List<StockSplit> getStockSplitsForSecurity(Security s) {
+    List<StockSplit> list = [];
+    for (StockSplit split in iterableList()) {
+      if (!s.isDeleted && split.security == s.uniqueId) {
+        list.add(split);
+      }
+    }
+    list.sort((final StockSplit a, final StockSplit b) {
+      return a.date!.compareTo(b.date!);
+    });
+
+    return list;
   }
 }